+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ChannelEditor.cs
