+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ChannelEditorType.cs
