+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ColorComponentEditor.cs
