+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ColorComponentTabViewController.cs
