+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ColorComponentViewController.cs
