+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ColorEditorLayer.cs
