+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ColorEditorView.cs
