+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/CommonBrushLayer.cs
