+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/CommonBrushView.cs
