+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/CommonGradientBrushLayer.cs
