+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ComponentSpinEditor.cs
