+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/HistoryLayer.cs
