+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/HueLayer.cs
