+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/MaterialBrushEditorViewController.cs
