+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/MaterialView.cs
