+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/ShadeLayer.cs
