+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/SolidColorBrushEditor.cs
