+++ Xamarin.PropertyEditing.Mac/Controls/Custom/CommonBrush/SolidColorBrushEditorViewController.cs
